move `APIAuthKey` type back to `api` package (#1765)
In order for there to be no source dependencies between `api` (which we intend to open source) and `server` (which we do not), I need to move the type definition for `APIAuthKey` back to the `api` package. I leave the `NewAPIAuthKeyFromRequest` function in the `server/auth` package, and turn the `Validate` and `AuthSystem` methods into functions that operate on an `APIAuthKey`.

fbshipit-source-id: 0256fc9