[fixes #877] CORS Lite; a minimal CORS impl (#2381)
This change adds config object embedded into a Domain that allows us to specify how OPTIONS (CORS) requests should be handled. This, in theory, will free up dealing with this in the application stack unless business logic needs to get involved.

One thing I'd like comment on is that I added nonstdlib/arrays/set. I'm not sure if this should be OSS or just a method somewhere. I didn't do a lot of work to make it particularly general / nice but it reads a little better than slice->map->etc when i want to compare two slices for order-ignoring equality.

Testing

Admittedly minimal for now. I've done all the unit testing and used tbnctl to test read/write path to DB.

To test routing I used the configwriter unit tests to produce a config block and pasted that into a test config that I was using on a local config.

fbshipit-source-id: 9daeb0b