[fixes #2369] Rewrite URL rewrite to enable a header constraint (#2431)
This adds the ability to require a test on a HTTP request's headers as part of rules controlling redirects.

Specifically the `Domain.Redirects.HeaderConstraint` struct was added that allows the following requirements to be added about a header:

- a specific header must be present
- a specific may not be present
- the value (optionally case sensitive or not) must be set on a specified header.

An example is given in the `Redirect` docs on how to use these new primitives to force SSL on endpoints.

At this time only one `HeaderConstraint` may be added though the structure supports an array for future evolution. If 2 or more are passed a validation error will be returned.

fbshipit-source-id: b407331