[fixes #2439] Significant rework of validity checking (#2489)
This change adds a more thorough / complete validation pass over objects being submitted. It also makes an effort to be consistent on how we represent a path to where an error was encountered. Work was done to provide an increase in the quality of tests covering the validity checking code paths.

As a related change this removed the deprecated `precreation` flag from the validity check.

fbshipit-source-id: 8c4eb23