[fixes #749][fixes #2535] Add support for domains to do SSL termination (#2536)
This adds experimental support for configuring our proxy to support SSL termination with the ability to specify: cert+key, (openssl) cipher filter, and supported protocols. The configuration is located within Domain.SSLConfig.

As a tiny added bonus this improves validity checking for whitespace-only values and fixes the validation path generated when a failure is found in domain aliases.

fbshipit-source-id: 6ff832c