[fixes #2547] Update NegativeMatch scope on ChangelogFilter (#2631)
This updates how a `changelog.Filter` is executed to exclude time from the portion that a negative match applies to. Specifically, with the example filter

```go
Filter{
  TimeRange: TimeRange{
    Start: threeHoursAgoUSec,
    End: nowUSec
  },
  NegativeMatch: true,
  Actor: "user-456",
  FieldFilter: FieldFilter {
    AttributePath: "cluster.instances%"
  }
}
```

Before this change would search for changes _not_ made by `user-456` AND were _not_ changes to a Cluster's Instance set AND were not in a trailing three hour window.

Now this same filter will search for changes that were _not_ made by `user-456` AND were _not_ changes to a Cluster's instance set but that _were_ in a trailing three hour window.

fbshipit-source-id: c68b162