[fixes #2546] Update HTTP handler to impose limits on public filter requests (#2702)
Prior to this change there was nothing implemented to ensure that the time ranges on each filter were set consistent with our API limitations. This updates the `History.Index` interface to take a global start/end time that will be used to check bounds for all filters. If a filter does not include an end cap it will be derived from the global start/end. If no start/end are provided (`time.Time{}`) then the default "now - 3 hours" will be used.

If any of the filters exceed the global window an error will be returned.

This additionally bumps the next oss minor version.

fbshipit-source-id: bdbf79f