Make fixtures stable
Previously we had fixtures exposed as a package variable but that meant it was easy to get in trouble by modifying the fixture which could impact downstream or even other package tests. Further if you needed a slight variation of the existing fixtures non trivial effort was required to copy the existing data so that changes to nested pointery attributes (slices, maps, pointers) did not impact both versions of the object.

This change replaces a single instance of a fixture with New() which will produce the same set of data that can the be modified (if necessary) and updates all calling code to use the new function.

fbshipit-source-id: 6f2a8a0