[fixes #2969] Updates API model to include ResponseData object
This adds a ResponseData object and links it to three other objects in decreasing specificity:

1. ClusterConstraint
2. Route
3. SharedRules

This change includes the IsValid and Equals implementation for the model updates and all related tests. It is the longest and most boring of the reviews. My guidance would be to start with response_data.go and look at how I've laid out the HeaderCookie data. Then check out Equality and Validity checks on the same. After that is comfortable work through ClusterConstraint (since it is nested within Routes and SharedRules) and finally tackle the remaining two objects.

fbshipit-source-id: 7d76afe