[api] introduce ProxyRef and ZoneRef (#4194)
This is another stop toward #4185. The goals of this change are:

1. Consolidate the way we turn Zone names into Zones, and Proxy/Zone names into Proxies
2. Make it possible to pass around a means to do these transformations. An envoy NodeID will be constructed from a Zone and Proxy name, and we will need to turn these NodeIDs into a proxy object from the API. It would be nice to defer the actual lookup into the parts of the code that are already dealing with the API.
3. Simplify flag-based configuration of these

To do this, we introduce `ProxyRef` and `ZoneRef`, which are built either from names or from objects. In the former case, the object can be subsequently populated by providing a `service.All`. In either case they expose a simple string that can be used for a map key, and which is derived from names, so it can be used even if the object hasn't been looked up yet.

We also restate `confagent.ProxyKeyFromFlags` as `api/client/flags/ProxyFromFlags` and `api/client/flags/ZoneKeyFromFlags` as `api/client/flags/ZoneFromFlags`, and reimplement them a flag-configured `ProxyRef` and `ZoneRef`, respectively.

Finally, we clean up all the code and tests, which was pretty automatic.

I plan to create `ProxyRefs` and `ZoneRefs` in the xds code from NodeIDs when creating Watches, and then from `*confagent.Objects` when consuming in the `RWObjectSource`.

fbshipit-source-id: 7e315be