[api] support setting of host header in the api client (#4311)
when you specify "Host" as a custom header in the flag-configured api client, it isn't properly respected. Digging into the way the go http library works, for client-side requests, the way to do this is to set Request.Host; If you set a "Host" header in a client-side request it is ignored. So this change looks for "host" headers and sets Request.Host appropriately. Behavior if the header is specified more than once is non-deterministic. I think that's fine.

fbshipit-source-id: 35bb609