[api] proxy refs have reversable map keys (#4629)
Our infra registers and deregisters with api.ProxyRefs; these have a MapKey() method, but have no obvious way to go from a map key string back to a ref.

Now that we need to be able to both marshal and unmarshal these keys, move the map keys to JSON representation, so they're easy to reverse. This also means we don't have to do any name escaping anymore, since the json encoder does this for us.

I explicitly panic on the case where the key can't be marshaled, as this should never ever happen.

fbshipit-source-id: d58d78a