[fixes #4753] Add Pagination to Changelog (#4873)
Update the structs returned by changelog endpoints to support pagination data.
Update the query processing to support reading pagination data from query args and documented in the API swagger.yml.

Minor version bump to 0.16.0 due to the update to the structure of envelope.Response and the responses from the changelog endpoints.

fbshipit-source-id: f3fb099