[fixes #5128] [fixes #4643] Fix inefficient Changelog query (#5265)
Adds:

- <Command>Ctx wrapping to our DB impl
- home-rolled long-query logging to our api-server (no stats for now, just a file)
- back pressure support for DB queries that exceed their deadline or get canceled by the DB
- Changelog DAL supports deadlines, 20s deadline added to all changelog requests
- Broke tbndb adapters into files per adapter for code hygiene
- Explicitly required tbndb objects to meet DB / Tx interfaces for safety
- Fixes a changelog query to not be absolutely terrible by not making the DB do join on 280k rows. This should resolve the RC of turbinelabs/incident#17

That last note is the only one that really matters here, everything else is iterative improvements that I made in the process of finding it. Runtime for some of the Nav queries reduces by up to ~95% (~35s -> 1.9s)

In the process of futzing with this I wrote the following test after getting a random suggestive error from lib/pq that led to the following program and me "verifying" that lib/pq will already handle cancelling queries that exceed some deadline.

fbshipit-source-id: 5f1c571