[rotor] make standalone differ product stable-ordered objects (#5891)
I noticed this while testing out the sidecar rotor-local.

When collected objects are written to the Houston API , the order of the objects (clusters, shared_rules, etc) is fixed and predictable. Unfortunately the objects are not reliably _collected_ in a stable order. The upshot is that when envoys pull config from a standalone rotor talking to kubernetes, a new config is sent every second, which isn't _terrible_ but also isn't great.

Rather than fix this in all collectors, we sort the instances and the clusters in the standalone differ. To do this we split the iteration over the diffs to first collect all the clusters, sorting the instances and metadata as we go, then sorting that cluster list, then producing the rest of the dependent objects.

Also fix a bug in api.InstancesByKey wherein instances with equal host values were incorrectly considered less, and fix some doc while we're in here.

fbshipit-source-id: d654b70